+++ source/renderer/app/actions/wallet-migration-actions.js
@@ -1,5 +1,6 @@
 // @flow
 import Action from './lib/Action';
+import type { ImportFromOption } from '../types/walletExportTypes';
 
 export default class WalletMigrationActions {
   startMigration: Action<any> = new Action();
@@ -8,5 +9,8 @@ export default class WalletMigrationActions {
   toggleWalletImportSelection: Action<{ index: number }> = new Action();
   updateWalletName: Action<{ index: number, name: string }> = new Action();
   nextStep: Action<any> = new Action();
-  selectExportSourcePath: Action<any> = new Action();
+  selectExportSourcePath: Action<{
+    importFrom: ImportFromOption,
+  }> = new Action();
+  resetExportSourcePath: Action<any> = new Action();
 }