+++ source/renderer/app/components/wallet/wallet-import/WalletImportFileDialog.js
@@ -11,11 +11,15 @@ import { Input } from 'react-polymorph/lib/components/Input';
 import { InputSkin } from 'react-polymorph/lib/skins/simple/InputSkin';
 import SVGInline from 'react-svg-inline';
 import classNames from 'classnames';
+import { get } from 'lodash';
 import styles from './WalletImportFileDialog.scss';
+import RadioSet from '../../widgets/RadioSet';
 import DialogCloseButton from '../../widgets/DialogCloseButton';
 import closeCrossThin from '../../../assets/images/close-cross-thin.inline.svg';
 import penIcon from '../../../assets/images/pen.inline.svg';
 import LoadingSpinner from '../../widgets/LoadingSpinner';
+import { ImportFromOptions } from '../../../types/walletExportTypes';
+import type { ImportFromOption } from '../../../types/walletExportTypes';
 
 const messages = defineMessages({
   title: {
@@ -35,17 +39,29 @@ const messages = defineMessages({
     defaultMessage: '!!!Select Daedalus state folder:',
     description: 'Select Daedalus state folder:',
   },
+  secretFileLabel: {
+    id: 'wallet.import.file.dialog.secretFileLabel',
+    defaultMessage: "!!!Select Daedalus 'secret.key' file:",
+    description: "Select Daedalus 'secret.key' file:",
+  },
   buttonLabel: {
     id: 'wallet.import.file.dialog.buttonLabel',
     defaultMessage: '!!!Import wallets',
     description: 'Import wallets',
   },
-  noWallets: {
-    id: 'wallet.import.file.dialog.noWallets',
+  stateDirNoWallets: {
+    id: 'wallet.import.file.dialog.stateDirNoWallets',
     defaultMessage:
       '!!!No wallets found. Make sure you have selected a Daedalus state directory which contains the ‘Secrets’ or `Secrets-1.0` folder with a `secret.key` file inside.',
     description:
       'No wallets found. Make sure you have selected a Daedalus state directory which contains the ‘Secrets’ or `Secrets-1.0` folder with a `secret.key` file inside.',
+  },
+  secretFileNoWallets: {
+    id: 'wallet.import.file.dialog.secretFileNoWallets',
+    defaultMessage:
+      '!!!No wallets found. Make sure you have selected a valid `secret.key` file.',
+    description:
+      'No wallets found. Make sure you have selected a valid `secret.key` file.',
   },
   linkLabel: {
     id: 'wallet.import.file.dialog.linkLabel',
@@ -58,6 +74,21 @@ const messages = defineMessages({
       '!!!https://iohk.zendesk.com/hc/en-us/articles/900000623463',
     description: '"Learn more" link URL on the wallet import file dialog',
   },
+  importFromLabel: {
+    id: 'wallet.import.file.dialog.importFromLabel',
+    defaultMessage: '!!!Import from:',
+    description: 'Import from:',
+  },
+  stateDirOptionLabel: {
+    id: 'wallet.import.file.dialog.stateDirOptionLabel',
+    defaultMessage: '!!!Daedalus state directory',
+    description: 'Daedalus state directory',
+  },
+  secretFileOptionLabel: {
+    id: 'wallet.import.file.dialog.secretFileOptionLabel',
+    defaultMessage: "!!!Daedalus 'secret.key' file",
+    description: "Daedalus 'secret.key' file",
+  },
 });
 
 type Props = {
@@ -68,24 +99,56 @@ type Props = {
   onClose: Function,
   onOpenExternalLink: Function,
   onSelectExportSourcePath: Function,
+  onResetExportSourcePath: Function,
   exportSourcePath: string,
+  defaultExportSourcePath: string,
+};
+
+type State = {
+  importFrom: ImportFromOption,
 };
 
 @observer
-export default class WalletImportFileDialog extends Component<Props> {
+export default class WalletImportFileDialog extends Component<Props, State> {
   static contextTypes = {
     intl: intlShape.isRequired,
   };
 
-  stateFolderInput: Input;
+  state = {
+    importFrom: ImportFromOptions.STATE_DIR,
+  };
+
+  importPathInput: Input;
 
   componentWillMount() {
     // Reset migration data
     this.props.onOpen();
   }
 
+  onSetImportFromOption = (importFrom: ImportFromOption) => {
+    if (this.state.importFrom !== importFrom) {
+      this.props.onResetExportSourcePath();
+      this.setState({ importFrom });
+    }
+  };
+
+  get input() {
+    const fallbackInput = {
+      blur: () => {},
+      focus: () => {},
+    };
+    return get(this, 'importPathInput.inputElement.current', fallbackInput);
+  }
+
+  isImportFromStateDir = (importFrom: ImportFromOption) =>
+    importFrom === ImportFromOptions.STATE_DIR;
+
+  isImportFromSecretFile = (importFrom: ImportFromOption) =>
+    importFrom === ImportFromOptions.SECRET_FILE;
+
   render() {
     const { intl } = this.context;
+    const { importFrom } = this.state;
     const {
       exportErrors,
       isSubmitting,
@@ -94,24 +157,26 @@ export default class WalletImportFileDialog extends Component<Props> {
       onOpenExternalLink,
       onSelectExportSourcePath,
       exportSourcePath,
+      defaultExportSourcePath,
     } = this.props;
+
     const title = intl.formatMessage(messages.title);
     const description = <FormattedHTMLMessage {...messages.description} />;
     const stateFolderLabel = intl.formatMessage(messages.stateFolderLabel);
+    const secretFileLabel = intl.formatMessage(messages.secretFileLabel);
     const buttonLabel = !isSubmitting ? (
       intl.formatMessage(messages.buttonLabel)
     ) : (
       <LoadingSpinner />
     );
     const linkLabel = intl.formatMessage(messages.linkLabel);
-    const noWalletError = intl.formatMessage(messages.noWallets);
+    const noWalletError = intl.formatMessage(
+      messages[`${importFrom}NoWallets`]
+    );
     const onLinkClick = () =>
       onOpenExternalLink(intl.formatMessage(messages.linkUrl));
 
-    const resetErrorCheck =
-      this.stateFolderInput &&
-      this.stateFolderInput.inputElement.current.value !== exportSourcePath;
-    const error = !resetErrorCheck && exportErrors !== '';
+    const error = exportErrors !== '';
 
     const inputClasses = classNames([
       styles.stateFolderInput,
@@ -119,7 +184,11 @@ export default class WalletImportFileDialog extends Component<Props> {
     ]);
 
     const buttonClasses = classNames(styles.actionButton, [
-      isSubmitting || error ? styles.disabled : null,
+      isSubmitting ||
+      error ||
+      (this.isImportFromSecretFile(importFrom) && !exportSourcePath)
+        ? styles.disabled
+        : null,
     ]);
 
     return (
@@ -137,43 +206,83 @@ export default class WalletImportFileDialog extends Component<Props> {
             icon={closeCrossThin}
             onClose={onClose}
           />
-          <div className={styles.backgroundContainer} />
           <div className={styles.content}>
             <div className={styles.title}>{title}</div>
             <div className={styles.description}>{description}</div>
+
+            <div className={styles.radioButtons}>
+              <RadioSet
+                label={intl.formatMessage(messages.importFromLabel)}
+                items={Object.keys(ImportFromOptions).map((key: string) => {
+                  const importFromOption: ImportFromOption =
+                    ImportFromOptions[key];
+                  return {
+                    key: importFromOption,
+                    label: intl.formatMessage(
+                      messages[`${importFromOption}OptionLabel`]
+                    ),
+                    selected: importFrom === importFromOption,
+                    onChange: () =>
+                      this.onSetImportFromOption(importFromOption),
+                  };
+                })}
+                verticallyAligned
+              />
+            </div>
+
             <div className={styles.stateFolderContainer}>
-              <p className={styles.stateFolderLabel}>{stateFolderLabel}</p>
+              <p className={styles.stateFolderLabel}>
+                {this.isImportFromStateDir(importFrom)
+                  ? stateFolderLabel
+                  : secretFileLabel}
+              </p>
               <div className={styles.stateFolderInputWrapper}>
                 <Input
                   type="text"
                   className={inputClasses}
                   ref={input => {
-                    this.stateFolderInput = input;
+                    this.importPathInput = input;
                   }}
                   skin={InputSkin}
-                  value={exportSourcePath}
+                  value={
+                    exportSourcePath ||
+                    (this.isImportFromStateDir(importFrom)
+                      ? defaultExportSourcePath
+                      : '')
+                  }
+                  placeholder={
+                    this.isImportFromStateDir(importFrom)
+                      ? defaultExportSourcePath
+                      : 'secret.key'
+                  }
+                  readOnly
                 />
                 <Button
                   className={styles.selectStateDirectoryButton}
-                  onClick={onSelectExportSourcePath}
+                  onClick={() => onSelectExportSourcePath({ importFrom })}
                   label={<SVGInline svg={penIcon} className={styles.penIcon} />}
                   skin={ButtonSkin}
                 />
               </div>
               {error && <p className={styles.noWalletError}>{noWalletError}</p>}
             </div>
+
             <div className={styles.action}>
               <Button
                 className={buttonClasses}
-                disabled={isSubmitting || error}
+                disabled={
+                  isSubmitting ||
+                  error ||
+                  (this.isImportFromSecretFile(importFrom) && !exportSourcePath)
+                }
                 label={buttonLabel}
                 onClick={onConfirm}
                 skin={ButtonSkin}
               />
             </div>
+
             <Link
               className={styles.learnMoreLink}
-              disabled={isSubmitting}
               onClick={onLinkClick}
               label={linkLabel}
               skin={LinkSkin}